Beatscrape_RegisteredUserLogin
Having trouble with getting registered users to log in. If they're credentials are correct, they're supposed to be redirected to the dashboard (hello.html) ... if incorrect, a message should be thrown that says 'Invalid username or password' ... instead, once entering the credentials (either correct or incorrect), nothing happens after hitting submitting.